// frontend/src/api/models/Bookmark.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Bookmark = {
    readonly id?: number;
    url: string;
    title?: string;
    description?: string;
    favicon_url?: string;
    readonly created_at?: string;
    readonly updated_at?: string;
};